Move TemplateStringToFile to fileutil
