+++ pkg/ddevapp/ddevapp.go
@@ -15,8 +15,6 @@ import (
 	"path/filepath"
 	"runtime"
 	"strconv"
-	"text/template"
-
 	"strings"
 
 	osexec "os/exec"
@@ -960,38 +958,10 @@ func (app *DdevApp) GenerateWebserverConfig() error {
 		content := string(c)
 		docroot := filepath.Join("/var/www/html", app.Docroot)
 
-		err = TemplateStringToFile(content, map[string]interface{}{"Docroot": docroot}, configPath)
+		err = fileutil.TemplateStringToFile(content, map[string]interface{}{"Docroot": docroot}, configPath)
 		if err != nil {
 			return err
 		}
-	}
-	return nil
-}
-
-// TemplateStringToFile takes a template string, runs templ.Execute on it, and writes it out to file
-func TemplateStringToFile(content string, vars map[string]interface{}, targetFilePath string) error {
-
-	templ := template.New("templateStringToFile:" + targetFilePath)
-	templ, err := templ.Parse(content)
-	if err != nil {
-		return err
-	}
-
-	var doc bytes.Buffer
-	err = templ.Execute(&doc, vars)
-	if err != nil {
-		return err
-	}
-
-	f, err := os.Create(targetFilePath)
-	if err != nil {
-		return err
-	}
-	defer util.CheckClose(f)
-
-	_, err = f.WriteString(doc.String())
-	if err != nil {
-		return nil
 	}
 	return nil
 }
